mc1n2: Allow swapping of microphone ADC channels
The SGH-I777 has the configuration for ADC0[0] and ADC0[1] swapped.
When using GT-I9100 libaudio, this results in the microphones
being improperly swapped between the main mic and the NR mic
near the earpiece.

This patch allows the mic channels to be swapped within the driver
to deal with incompatible libaudio.

It is configurable in the event that SGH-I777 libaudio becomes
available.

Signed-off-by: Andrew Dodd <atd7@cornell.edu>